+++ src/main/java/com/codetaylor/mc/pyrotech/modules/plugin/dropt/ModulePluginDroptConfig.java
@@ -13,7 +13,6 @@ public class ModulePluginDroptConfig {
   static {
     String[] rules = {
         "grass_tall",
-        "leaves",
         "dirt",
         "grass",
         "sand",